small file change, seperated Choice_Menu into seperate file. Although other choice buttons had errors such as 'circular import', and 'module not found'
